alphagEqn: Tesing different things
Test by activating the source term by h90 (or hi). Seemed to be a litle
bit probmematic since positive dUdn-values did not correspond to h90.
Therefore, dUdn=const - just for testing. This tended to result in too
much air within the lower parts of the spillway, which totally blowed up,
basically because of a too wide h90-line->region. To smarten this up, a
distance cut-off was introduced, so that no air should be introduced below
0.5(?)*h100. This was basically to avoid activation of the sourceterm
within the steps.